MAGE-1320 Relocate price key resolver
